+++ History.md
@@ -13,6 +13,7 @@
   - Votings can be chosen at the creation wizard of a site or consultation.
 - It is now logged when users are added to or removed from a user group. A log is visible for admins.
 - It is now possible for an admin to show the last edit date of a motion or amendment on the home page.
+- Reading comments can now be restricted to the same user group as writing comments.
 - Bugfix: Base statute texts were shown in the "New motions" section in the sidebar.
 
 ### Version 4.10.1 (2022-07-01)